+++ src-input/duk_internal.h
@@ -67,6 +67,7 @@ DUK_USE_USER_DECLARE()
 #include "duk_harray.h"
 #include "duk_henv.h"
 #include "duk_hbuffer.h"
+#include "duk_hproxy.h"
 #include "duk_heap.h"
 #include "duk_debugger.h"
 #include "duk_debug.h"